Day1

+++ Uebung00/Aufgabe1.py
@@ -1,8 +1,8 @@
 firstNumber = input("Enter your first Number: ")
 operator = input("Enter your Operator(+,-,/,*): ")
 secondNumber = input("Enter your second Number: ")
-firstNumber = int(firstNumber)
-secondNumber = int(secondNumber)
+firstNumber = float(firstNumber)
+secondNumber = float(secondNumber)
 if operator == "+":
     print(firstNumber + secondNumber)
 elif operator == "-":
